Threaded control/render with buffered variables
Polling the GLFW events and controlling the logic of the program runs in
one thread (called the control thread), and the rendering runs in
thread called the render thread.
To support this, I've seperated the variables into two groups.
One called "Unique Variables",
which is two buffer variables that can be directly changed each loop
of the control thread, so it swapped each loop of the control thread and
only copied to a third buffer of the render thread at the start of the
loop of the render thread (while maintaining synchronization using a
mutex synchronization primitive).
The other group of variables is called "Shared Variables" and it is only
buffered once (a signle copy), and these variables are only changed
in the control thread through "pending" methods (e.g. resize_pending
in ScreenPixels) and are sync'd at the start of the render thread loop
(the pending calls and the sync calls are synchronized through
in-class mutexes).