Fix crashing (rendering broken; AntTweakBar broken)
Although AntTweakBar doesn't render, I can still load models by
pressing INSERT on the keyboard.
But the rendering is kind of broken too, the lighting is not as
it's suppose to be. That's probably because of the changes I made to
the lighting and/or the move to Vector3 in many places.